+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Minecarts and Boats are now easier to break
+
 ## 3.9.7
 * Increased Flint and Steel durability to 61
 * Fixed crash on world creation when missing Supplementaries