+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Minecarts and Boats are now easier to break
 * Sea lanterns now break faster with pickaxes
+* Reduced enchanting levels for fished Fishing Rods and Bows
 * Fixed grave not generating if wearing only a tool belt
 * Fixed Item Stats and stack sizes applied on `/reload` instead of post reload, conflicting